+++ packages/interact/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @glissade/interact
 
+## 0.63.2-pre.1
+
+### Patch Changes
+
+- @glissade/core@0.63.2-pre.1
+- @glissade/player@0.63.2-pre.1
+- @glissade/scene@0.63.2-pre.1
+
 ## 0.63.2-pre.0
 
 ### Patch Changes