+++ packages/scene/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @glissade/scene
 
+## 0.63.2-pre.1
+
+### Patch Changes
+
+- @glissade/core@0.63.2-pre.1
+
 ## 0.63.2-pre.0
 
 ### Patch Changes